Use local font classes to avoid build font downloads
